przyjmowanie zamowien przez web socketa

+++ src/main/java/com/foodapp/foodapp/company/CompanyRepository.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 @Repository
 public interface CompanyRepository extends JpaRepository<Company, Long> {
-    Optional<Company> findById(Integer companyId);
+    Optional<Company> findById(Long companyId);
 }
